use rescaling to get continuous pixel sizes

+++ EyeTrackingSuite/ETSProsthesis.h
@@ -6,7 +6,7 @@ typedef struct
 {
 	bool changed;
 	int grayLevels;
-	int pixelSize;
+	double pixelSize;
 	int fullBlack;
 	int fullWhite;
 } ETSProsthesisDrawOptions;
